pyproject.toml
[project]
name = "booktracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.135.2,<0.136.0)",
    "streamlit (>=1.55.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (==4.0.1)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
test_lint = [
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "locust (>=2.43.3,<3.0.0)",
    "radon (>=6.0.1,<7.0.0)",
    "bandit (>=1.9.4,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]
